% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhistogram.R
\name{bhistogram}
\alias{bhistogram}
\title{Return a binned histogram of array or vector}
\usage{
bhistogram(v, title = NULL, title_size = 24, n_breaks = 10)
}
\arguments{
\item{v}{numeric}

\item{title}{character}

\item{title_size}{numeric}

\item{n_breaks}{numeric}
}
\value{
ggplot object
}
\description{
Create binned histogram for numeric data
}
\examples{
v = c(rnorm(1500, 0, .1), rnorm(1000, 1, .2))
bhistogram(v)
}
