% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fat_192x96x1}
\alias{fat_192x96x1}
\title{Fat 192x96x1}
\format{
A list of tensors. Images resolution 192x96.
\describe{
\item{train}{number of samples 102}
\item{test}{number of samples 10}
}
}
\usage{
fat_192x96x1
}
\description{
Fat 192x96x1
}
\keyword{datasets}
