% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fat_384x192x1}
\alias{fat_384x192x1}
\title{Fat 384x192x1}
\format{
A list of tensors. Images resolution 384x192.
\describe{
\item{train}{number of samples 67}
\item{test}{number of samples 9}
}
}
\usage{
fat_384x192x1
}
\description{
Fat 384x192x1
}
\keyword{datasets}
