% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fat_768x384x1}
\alias{fat_768x384x1}
\title{Tensors for U-Net Neural Nets}
\format{
A list of tensors. Images resolution 768x384.
\describe{
\item{train}{number of samples 67}
\item{test}{number of samples 9}
}
}
\usage{
fat_768x384x1
}
\description{
Tensors were created from 80 768x384 images
of the 80 patients,
along with manually created masks for
subcutaneous adipose tissue (SCAT)
and visceral adipose tissue (VSAT).
}
\details{
Tensors are arranged into two lists,
train and test list, respectively.
There are eight tensors:
train$image, train$mask, train$scat, train$vsat,
test$image, test$mask, test$scat, test$vsat.

Each tensor is an array of image data
with shape: train -- (67, 384, 192), test -- (67, 384, 192)
Fat 768x384x1
}
\keyword{datasets}
