% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_info.R
\name{images_info}
\alias{images_info}
\title{Return a list of five 2D arrays}
\usage{
images_info(images_dir, extension = "dcm")
}
\arguments{
\item{images_dir}{character}

\item{extension}{character}
}
\value{
tibble
}
\description{
Create five named 2D arrays: checker, urandom, zeros, ones, rectangle.
}
\examples{
\dontrun{
images_info("../80_images", extension = "dcm")
}
}
