% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_arrays.R
\name{make_arrays}
\alias{make_arrays}
\title{Return a list of five 2D arrays}
\usage{
make_arrays(m = 5, n = 5)
}
\arguments{
\item{m}{numeric}

\item{n}{numeric}
}
\value{
list of five named 2D arrays
}
\description{
Create five named 2D arrays: checker, urandom, zeros, ones, rectangle.
}
\examples{
make_arrays(5, 5)
}
