% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_array2d.R
\name{plot2_array2d}
\alias{plot2_array2d}
\title{Create ggplot from array and mask}
\usage{
plot2_array2d(
  arrList,
  maskList = NULL,
  alpha = 0.35,
  title = NULL,
  title_size = 18
)
}
\arguments{
\item{arrList}{array or list of 1, 2, or 3 arrays}

\item{maskList}{array or list of 1, 2, or 3 arrays}

\item{alpha}{numeric}

\item{title}{character}

\item{title_size}{numeric}
}
\value{
ggplot object
}
\description{
Create ggplot from array and mask
}
\examples{
urandom <- make_arrays(m = 5, n = 5)$urandom
square <- make_arrays(m = 5, n = 5)$rectangle
plot2_array2d(urandom, square, title = "urandom*square", title_size = 18)
}
