% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_array2d.R
\name{plot_array2d}
\alias{plot_array2d}
\title{Return a ggplot of array}
\usage{
plot_array2d(arr, title = NULL)
}
\arguments{
\item{arr}{array}

\item{title}{character}
}
\value{
ggplot object
}
\description{
Create ggplot from array
}
\examples{
urandom <- make_arrays(m = 5, n = 5)$urandom
plot_array2d(urandom)
}
