% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_array2d.R
\name{plotBlendedImages}
\alias{plotBlendedImages}
\title{Create ggplot from image and mask}
\usage{
plotBlendedImages(img, mask, alpha = 0.35, title = NULL, title_size = 16)
}
\arguments{
\item{img}{antsImage}

\item{mask}{antsImage}

\item{alpha}{numeric}

\item{title}{character}

\item{title_size}{numeric}
}
\value{
ggplot object
}
\description{
Create ggplot from image and mask
}
