% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_array2d.R
\name{plot_array2d}
\alias{plot_array2d}
\alias{plotBlendedImages}
\title{Return a ggplot of array}
\usage{
plot_array2d(arrList, title = NULL, title_size = 18)

plotBlendedImages(img, mask, alpha = 0.35)
}
\arguments{
\item{arrList}{array or list of 1, 2, or 3 arrays}

\item{title}{character}

\item{title_size}{numeric}

\item{img}{antsImage}

\item{mask}{antsImage}

\item{alpha}{numeric}
}
\value{
ggplot object

ggplot object
}
\description{
Create ggplot from array
}
\examples{
urandom <- make_arrays(m = 5, n = 5)$urandom
plot_array2d(urandom, title = "urandom", title_size = 24)
}
